+++ frontend/src/App.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import SignupFormPage from "./components/SignupFormPage";
-import LoginFormPage from "./components/LoginFormPage";
+import SignupForm from "./components/SignupForm";
+import LoginForm from "./components/LoginForm";
 import Layout from "./components/Layout";
 
 function App() {
   return (
     <Layout>
       <Routes>
-        <Route path="/login" element={<LoginFormPage />} />
-        <Route path="/signup" element={<SignupFormPage />} />
+        <Route path="/login" element={<LoginForm />} />
+        <Route path="/signup" element={<SignupForm />} />
       </Routes>
     </Layout>
   );
